[FIX] server_ftp. Correct super() call.
Also refactor make MockServerCase class available for custom tests.
