+++ server_ftp/tests/__init__.py
@@ -1,2 +1,3 @@
 # License AGPL-3.0 or later (https://www.gnu.org/licenses/agpl.html).
+from . import common
 from . import test_mock_server
